빌드 플래그 수정 , 16MB 디버그 개발 환경 오류 수정

+++ lib/MUFFIN/src/Protocol/SPEAR/SPEAR.cpp
@@ -869,7 +869,7 @@ namespace muffin {
         JsonArray value = response["7"].as<JsonArray>();
         for (auto val : value)
         {
-            daq->PolledValuesVector.emplace_back(val.as<uint16_t>());
+            daq->PolledValuesVector.emplace_back(val.as<int32_t>());
         }
         xSemaphoreGive(xSemaphoreSPEAR);
         return Status(Status::Code::GOOD);
